make new level 2, change level to float

+++ TaraV1/TaraV1/Assets/Scripts/ColliderPickup.cs
@@ -47,7 +47,7 @@ public class ColliderPickup : MonoBehaviour
     public int loadingCutsceneTo;
     public int cutscene_duration = 3;
 
-    public int loadingLevel;
+    public float loadingLevel;
 
 	public bool Active{
 		set {
